% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psData-methods.R
\name{add}
\alias{add}
\title{Add data to a psData object}
\usage{
add(x, newData)
}
\arguments{
\item{x}{an object of class \code{psData}---see \code{\link{readData}} for details.}

\item{newData}{either a \code{vector}, \code{matrix} or \code{data.frame}
containing the new data. If a \code{vector} or \code{magtrix} is supplied then it must
be either of length or have two columns. If a \code{data.frame} is supplied then the columns
must be labelled \code{"n"} and \code{"rn"}. The new data MUST NOT contain values that already
exist in \code{x$n}}
}
\value{
an object of class \code{pSData}
}
\description{
Add one or more new observations to an existing clothing survey object.
}
\examples{
add(Ssurveys$lau, c(11, 1))
}
