% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootFit.R
\name{bootFit}
\alias{bootFit}
\title{importFrom doParallel regusterDoParallel
import foreach
importFrom pbapply pblapply pbsapply pboptions
importFrom parallel detectCores makeCluster parApply parLapply parSapply stopCluster}
\usage{
bootFit(
  x,
  B = 2000,
  model = c("zeta", "zi.zeta"),
  silent = FALSE,
  parallel = TRUE,
  progressBar = FALSE,
  pbopts = list(type = "txt")
)
}
\description{
importFrom doParallel regusterDoParallel
import foreach
importFrom pbapply pblapply pbsapply pboptions
importFrom parallel detectCores makeCluster parApply parLapply parSapply stopCluster
}
