% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.psFit.R
\name{confint.psFit}
\alias{confint.psFit}
\title{S3 confint method for objects of class psFit}
\usage{
\method{confint}{psFit}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{psFit} returned by \code{fit()}.}

\item{parm}{added for compatibility. Should be left empty as it is ignored.}

\item{level}{the confidence level required---restricted to [0.75, 1)}

\item{...}{in theory other parameters to be passed to \code{confint}, but in
reality passed as extra parameters to the internal function \code{plZIZ}.}
}
\value{
if the zeta model is used (i.e \code{object} comes from a call to
  \code{fit(..., model = zetaModel())}), then a list with two items: \code{wald} and
  \code{prof} containing the Wald and profile likelihood confidence intervals
  respectively for the shape parameter of the fitted zeta distribution is
  returned. In general these should be relatively close to each other. These
  values use the zeta distribution shape parameter and must satisfy
  \code{shape > 1}. If a
  zero-inflated zeta model is used then a list of a confidence regions is returned with
  and element for each value of \code{level}. The confidence regions are
  \code{data.frame}s with variables \code{pi} and \code{shape} which can be
  used with \code{\link[graphics]{lines}} or \code{\link[graphics]{polygon}}
  to draw a the confidence region.
}
\description{
S3 confint method for objects of class psFit
}
\details{
NOTE: the method for ZIZ model is a little computationally intensive
and possibly (almost certainly) unstable.
}
\examples{
data(Psurveys)
roux = Psurveys$roux
mleFit = fit(roux, model = zetaModel())
confint(mleFit)

if (interactive()) {
  zizFit = fit(roux, model = zizModel())
  confint(zizFit, level = c(0.80, 0.95))
  plotUncertainty(zizFit, level = c(0.80, 0.95))
}

}
