% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.psData.R
\name{fitted.psData}
\alias{fitted.psData}
\title{S3 fitted method for an object of class \code{psFit}}
\usage{
\method{fitted}{psData}(object, ...)
}
\arguments{
\item{object}{an object of class \code{psFit}, usually from \code{\link{fitDist}}
or \code{\link{fitZIDist}}.}

\item{...}{other arguments passed to \code{fitted}f---not used.}
}
\value{
a named vector of fitted probabilities
}
\description{
S3 fitted method for an object of class \code{psFit}
}
