% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psPosterior.R
\name{fitted.psPosterior}
\alias{fitted.psPosterior}
\title{Extract posterior mean probabilities}
\usage{
\method{fitted}{psPosterior}(object, ...)
}
\arguments{
\item{object}{An object of class `psPosterior`.}

\item{...}{Additional arguments retained for S3 compatibility.}
}
\value{
A named numeric vector of posterior mean probabilities.
}
\description{
Extract posterior mean probabilities
}
