% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psData-methods.R
\name{mean.psData}
\alias{mean.psData}
\title{An S3 method for computing the mean of clothing survey for the number of
groups or size of groups}
\usage{
\method{mean}{psData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{psData}---\code{\link{readData}} for more
details.}

\item{...}{other arguments which are passed to \code{\link[base]{sum}}}
}
\value{
the mean of the data. If there are \eqn{r_i}{r[i]} observations of
  the value \eqn{n_i}{n[i]} then the mean is given by
  \deqn{\sum_i\frac{r_i\times n_i}{\sum_i{r_i}}}{sum(r[i]*n[i])/sum(r[i])}.
}
\description{
An S3 method for computing the mean of clothing survey for the number of
groups or size of groups
}
\examples{
data(Psurveys)
mean(Psurveys$roux)
}
