% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psModel.R
\name{modelParameterNames}
\alias{modelParameterNames}
\title{Return the natural parameter names declared by a fitPS model.}
\usage{
modelParameterNames(model, ...)
}
\arguments{
\item{model}{A `psModel` object.}

\item{...}{Additional arguments reserved for model methods.}
}
\value{
Character vector of parameter names.
}
\description{
Return the natural parameter names declared by a fitPS model.
}
