% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psFit.R
\name{plot.psFit}
\alias{plot.psFit}
\title{S3 plot method for an object of class \code{psFit}}
\usage{
\method{plot}{psFit}(x, ylim = c(0, 1), ...)
}
\arguments{
\item{x}{an object of class \code{psFit}, usually from \code{\link{fitDist}}}

\item{ylim}{the limits of the y-axis}

\item{...}{other arguments passed to \code{plot}}
}
\value{
No return value, called for side effects
}
\description{
S3 plot method for an object of class \code{psFit}
}
\examples{
p = readData(system.file("extdata", "p.xlsx", package = "fitPS"))
fit = fitDist(p)
plot(fit)
}
