% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorProbs.R
\name{posteriorProbs}
\alias{posteriorProbs}
\alias{posteriorProbs.psFit}
\alias{posteriorProbs.psPosterior}
\title{Extract posterior probability summaries}
\usage{
posteriorProbs(object, ...)

\method{posteriorProbs}{psFit}(object, n = NULL, ...)

\method{posteriorProbs}{psPosterior}(object, n = NULL, ...)
}
\arguments{
\item{object}{A Bayesian `psFit` object or a `psPosterior` object.}

\item{...}{Additional arguments passed to methods.}

\item{n}{`NULL`, a single number of leading probabilities to return, or a
vector of P or S indices to return.}
}
\value{
A data frame containing posterior probability summaries.
}
\description{
Extract posterior summaries for fitted P or S probabilities from a Bayesian
fitPS model or from its associated posterior object.
}
