% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psData-methods.R
\name{print.psData}
\alias{print.psData}
\title{S3 print method for an object of class \code{psData}}
\usage{
\method{print}{psData}(x, ...)
}
\arguments{
\item{x}{an object of class \code{psData}, usually from \code{\link{readData}}
or \code{\link{makePSData}}}

\item{...}{other arguments passed to \code{print}}
}
\value{
No return value, called for side effects
}
\description{
S3 print method for an object of class \code{psData}
}
