% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psFit-methods.R
\name{print.psFit}
\alias{print.psFit}
\title{S3 print method for an object of class \code{psFit}}
\usage{
\method{print}{psFit}(x, nterms = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{psFit}, usually returned by [fit()] or one
of the deprecated compatibility fitters.}

\item{nterms}{Number of probability terms to print. If \code{NULL}, print
the terms already stored in the fitted object, capped at 10 for posterior
and bootstrap summaries.}

\item{...}{Other arguments passed to delegated print methods.}
}
\value{
No return value, called for side effects.
}
\description{
S3 print method for an object of class \code{psFit}
}
