% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psPosterior.R
\name{print.psPosterior}
\alias{print.psPosterior}
\title{Print a fitPS posterior object}
\usage{
\method{print}{psPosterior}(x, ...)
}
\arguments{
\item{x}{An object of class `psPosterior`.}

\item{...}{Additional arguments passed to `print.data.frame()`.}
}
\value{
The posterior object, invisibly.
}
\description{
Print a fitPS posterior object
}
