% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBootstrap.R
\name{summary.psBootstrap}
\alias{summary.psBootstrap}
\title{Summarise a fitPS bootstrap object}
\usage{
\method{summary}{psBootstrap}(object, ...)
}
\arguments{
\item{object}{An object of class `psBootstrap`.}

\item{...}{Additional arguments retained for S3 compatibility.}
}
\value{
An object of class `summary.psBootstrap`.
}
\description{
Summarise a fitPS bootstrap object
}
