% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psPosterior.R
\name{summary.psPosterior}
\alias{summary.psPosterior}
\title{Summarise a fitPS posterior object}
\usage{
\method{summary}{psPosterior}(object, ...)
}
\arguments{
\item{object}{An object of class `psPosterior`.}

\item{...}{Additional arguments retained for S3 compatibility.}
}
\value{
An object of class `summary.psPosterior`.
}
\description{
Summarise a fitPS posterior object
}
