% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psModel.R
\name{supportedPosteriorEngines}
\alias{supportedPosteriorEngines}
\title{Return posterior engines supported by a fitPS model.}
\usage{
supportedPosteriorEngines(model, ...)
}
\arguments{
\item{model}{A `psModel` object.}

\item{...}{Additional arguments reserved for model methods.}
}
\value{
Character vector of posterior-engine identifiers.
}
\description{
Return posterior engines supported by a fitPS model.
}
