% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCI.R
\name{bootCI}
\alias{bootCI}
\alias{bootCI.psData}
\alias{bootCI.psFit}
\title{Bootstrap confidence intervals or regions}
\usage{
bootCI(x, ...)

\method{bootCI}{psData}(x, ...)

\method{bootCI}{psFit}(x, ...)
}
\arguments{
\item{x}{a object of class \code{psFit}---see \code{\link{readData}} for more
details.}

\item{level}{the confidence level required---restricted to [0.75, 1). This
may be a vector, in which case multiple intervals, or confidence regions
will be returned.}

\item{B}{the number of bootstrap samples to take.}

\item{model}{which model to fit to the data, either \code{"zeta"} or
\code{"zi.zeta"}. Maybe abbreviated to \code{"z"} and \code{"zi"}. Default
is \code{"zeta"}.}

\item{silent}{if \code{TRUE} then no output will be displayed whilst the
bootstrapping is being undertaken. \code{plot} if \code{TRUE} then the
contours for the confidence region will be plotted. This only works if
\code{model = "zi.zeta"}. It is ignored otherwise. \code{parallel} if
\code{TRUE} then the bootstrapping is performed in parallel. This is
currently \code{FALSE} by default because, for reasons I do not understand,
the parallel code is slower than the sequential code.}

\item{progressBar}{if \code{TRUE} then progress bars will be displayed to
show progress on the bootstrapping.}

\item{pbopts}{a list of arguments for the \code{\link[pbapply]{pboptions}}
function that affect the progress bars. Ignored if \code{progressBar =
FALSE}.}
}
\description{
Use boostrapping to generate confidence intervals, or confidence regions in
the case of the zero-inflated model.
}
\details{
This function uses bootstrapping to compute a confidence interval
  for the shape parameter in the case of the zeta model and a confidence
  region in the case of the zero-inflated zeta model. A smoothed bootstrap
  approach is taken rather than a simple percentile method. The kernel
  density estimation is performed by the \code{ks} package using a smoothed
  cross-validated bandwidth selection procedure. WARNING: although this
  function does offer parallel computation, at this point in time it is
  slower that sequential evaluation. Hopefully this will not be the case in the
  future.
}
\section{Methods (by class)}{
\itemize{
\item \code{bootCI(psData)}: Bootstrap confidence intervals or regions

\item \code{bootCI(psFit)}: Bootstrap confidence intervals or regions

}}
\examples{
\dontrun{
data(Psurveys)
roux = Psurveys$roux
confRegion = bootCI(roux, model = "zi.zeta", parallel = FALSE, plot = TRUE)

## This will not work unless you have the sp package installed
## Count how many of the points lie within the 95\% confidence region
table(sp::point.in.polygon(fit$pi,fit$shape, confRegion$pi, confRegion$shape))
}
}
