% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootFit.R
\name{bootFit}
\alias{bootFit}
\title{importFrom pbapply pblapply pbsapply}
\usage{
bootFit(x, B = 2000, model = c("zeta", "zi.zeta"), progress.bar = FALSE)
}
\description{
importFrom pbapply pblapply pbsapply
}
