% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPSData.R
\name{createPSData}
\alias{createPSData}
\title{Manually create psData}
\usage{
createPSData(n, rn, type = c("P", "S"), notes = NULL)
}
\arguments{
\item{n}{a vector of lables for the number of groups or size of groups of
glass.}

\item{rn}{a vector of counts corresponding to each element of \code{n}. All
entries must be greater than zero.}

\item{type}{either \code{"P"} or \code{"S"}.}

\item{notes}{a character string or a \code{\link[utils]{bibentry}}.}
}
\value{
an object of class \code{psData}. See \code{\link{readData}} for more
  details
}
\description{
A mechanism for manually creating P or S data sets for use with fitDist.
}
\examples{
p = createData(0:2, c(98, 1, 1), type = "P")
p
}
