% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credint.R
\name{credint}
\alias{credint}
\title{Bayesian credible intervals or regions}
\usage{
credint(psFit, level = 0.95, plot = FALSE, silent = FALSE, ...)
}
\arguments{
\item{psFit}{a object of class \code{psFit}.}

\item{level}{the credible level required---restricted to [0.75, 1). This may
be a vector, in which case multiple intervals, or credible regions will be
returned.}

\item{plot}{if \code{TRUE} and \code{model == "ziz"}, then a plot of the
bootstrapped values will be produced and confidence contour lines will be
drawn for each value in level.}

\item{silent}{if \code{TRUE}, then no output will be displayed whilst the the
kernel density estimation is being undertaken.}

\item{...}{other arguments fed to plot. If \code{plot == FALSE}, then these
will be ignored}
}
\value{
If \code{psData$model == "zeta"}, then either a \code{vector} or a \code{data.frame}
 with elements/columns named \code{"lower"} and \code{"upper"} representing
 the lower and upper bounds of the confidence interval(s). Multiple bounds
 are returned in a \code{data.frame} when \code{level} has more than one
 value. If \code{psData$model == "ziz"}, then a list with length equal to the
 length of \code{level} is returned. The name of each element in the list is
 the level with % attached. For example if \code{level == 0.95}, then the
 list has a single element named \code{"95\%"}. It is possible for there to
 be multiple contours for the confidence region for a given \code{level}. If
 there is only one contour for each value of \code{level}, then each element
 of the list consists of a \code{list} with elements named \code{pi} and
 \code{shape} which specify the coordinates of the contour(s) for that level.
 There is a third element named \code{level} which gives the height of the
 kernel density estimate at that contour. If there are multiple contours for
 a given value of \code{level} then each list element is a list of lists with
 the structure given above (\code{level}, \code{pi}, and \code{shape}). NOTE:
 it is quite possible that there are multiple contours for a given height. If
 you want a way of thinking about this consider a mountain range with two
 mountains of equal height. If you draw the contours for (almost) any
 elevation, then you would expect to capture a region from each mountain.
}
\description{
Use kernel density estimation to generate credible intervals, or credible
regions in the case of the zero-inflated model.
}
\details{
This function uses kernel density estimation to compute a Bayesian
 credible interval for the shape parameter in the case of the zeta model and
 a credible region in the case of the zero-inflated zeta model. A smoothing
 approach is taken rather than a simple percentile method. The
 kernel density estimation is performed by the \code{ks} package using a
 smoothed cross-validated bandwidth selection procedure.
}
\examples{
\dontrun{
data(Psurveys)
roux = Psurveys$roux
fit = fitzidist(roux, method == "bayes")
credRegion = credint(roux, plot = TRUE)

## This will not work unless you have the sp package installed
## Count how many of the points lie within the 95\% confidence region
lapply(credRegion, function(cr){
  table(sp::point.in.polygon(fit$pi,fit$shape, cr$pi, cr$shape))
. })
}
}
