% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.psFit.R
\name{fitted.psFit}
\alias{fitted.psFit}
\title{S3 fitted method for an object of class \code{psFit}}
\usage{
\method{fitted}{psFit}(object, n = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{psFit}, usually from \code{\link{fitDist}}
or \code{\link{fitZIDist}}.}

\item{n}{This parameter is \code{NULL} by default. If it is not \code{NULL}
 then it must be either the number of fitted terms to be return, or, a vector
containing the desired fitted values.}

\item{...}{other arguments passed to \code{fitted}---not used.}
}
\value{
a named vector of fitted probabilities
}
\description{
S3 fitted method for an object of class \code{psFit}
}
