% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitZIDist.R
\name{fitZIDist}
\alias{fitZIDist}
\alias{fitZIdist}
\alias{fitzidist}
\title{Fit a Zero-Inflated Zeta Distribution to Forensic Data}
\usage{
fitZIDist(x, nterms = 10, method = c("mle", "bayes"), ...)

fitZIdist(x, nterms = 10, method = c("mle", "bayes"), ...)

fitzidist(x, nterms = 10, method = c("mle", "bayes"), ...)
}
\arguments{
\item{x}{an object of type \code{psData}, usually obtained from
\code{\link{readData}}.}

\item{nterms}{the number of terms to compute the probability distribution
for.}

\item{method}{either \code{"mle"} or \code{"bayes"}. Lets the user choose
maximum likelihood estimation or Bayesian estimation. NOTE: each of these
modes of estimation has a different set of optional parameters and
defaults. See the description of the \code{\ldots} parameter below for
details.}

\item{...}{other arguments that control the estimation methods. If
  \code{method == "mle"}, then the user can provide an optional argument
  \code{start} which is the starting value for the numerical optimisation. If
  this is not provided, then \code{start = c(0.5, 1)} by default. If you specify your
  own starting value, it would be sensible to keep it above 0.5 for pi and 1 for the shape.

  If \code{method == "bayes"}, then there are seven optional parameters (which
  despite the documentation are actually case insensitive):
\describe{
  \item{\code{theta0}}{ -- The initial value of the mixing parameter and the shape parameter, set to \code{c(0.5, 1), by default} }
  \item{\code{a}}{ -- The lower bound of the limit for the uniform distribution for the shape parameter prior which is U[a,b] .The default is -2. }
  \item{\code{b}}{ -- The upper bound of the limit for the uniform distribution for the shape parameter prior which is U[a,b] .The default is +2. }
  \item{\code{alpha}}{ -- The first shape parameter for the beta prior on the mixing distribution which is Beta(alpha, beta) .The default is 1. }
  \item{\code{beta}}{ -- The second shape parameter for the beta prior on the mixing distribution which is Beta(alpha, beta) .The default is 1. }
  \item{\code{nIter}}{ -- The number of samples to save from the chain. Must be greater than zero, and ideally greater than 1000. }
  \item{\code{nBurnIn}}{ -- The number of samples to discard from the chain. Must be greater than zero. **NOTE**: the sampler runs for \code{nIter + nBurnIn} iterations,
  so you do not need to factor this number into your number of samples, \code{nIter}. }
  \item{\code{silent}}{ -- A logical variable which allows the user to get a progress bar if they want. \code{TRUE} by default. }
}}
}
\value{
an object of class \code{psFit}--see Details.
}
\description{
This function uses maximum likelihood estimation (MLE) or Bayesian estimation
(MCMC) to estimate the mixing parameter and the shape parameter of a
zero-inflated zeta distribution from a set of observed counts for either the
number of groups/sources of forensically interesting material (mostly glass
or paint) recovered from clothing, or the number of fragments/particles in
each group. This, in turn, allows the estimation of the P and S
probabilities, as described by Evett and Buckleton (1990), which used in
computing the likelihood ratio (LR) for activity level propositions. The data
itself arises from clothing surveys. The zero-inflated zeta distribution has
probability mass function
\deqn{p(k) = \begin{cases}
\pi + \frac{(1-\pi)}{\zeta(s)}&,k=0, \\
\frac{(1-\pi)k^{-s}}{\zeta(s)}&,k=1,2,\ldots
\end{cases}
}{
p(0) = pi + (1-pi)/zeta(s),k = 0 and p(k) = (1-pi)k^-s/zeta(s),k>1
}
where \eqn{\zeta(s)}{zeta(s)} is the Reimann Zeta function.
}
\details{
The function returns an object of class \code{psFit} which is a
  \code{list} contains eight or nine elements:
\describe{
\item{\code{psData}}{ -- an object of class \code{psData}--see \code{\link{readData}},}
\item{\code{fit}}{ -- the fitted object from \code{\link[stats]{optim}},}
\item{\code{pi}}{ - the maximum likelihood estimate, or the posterior mean, of the mixing parameter,}
\item{\code{shape}}{ -- the maximum likelihood estimate, or the posterior mean, of the shape parameter,}
\item{\code{var.cov}}{ -- the estimated (posterior) variance-covariance matrix for the parameters,}
\item{\code{fitted}}{ -- a named \code{vector} containing the first \code{nterms} of the fitted distribution.}
\item{\code{model}}{ -- set to \code{"ziz"} for this model,}
\item{\code{method}}{ -- the method of estimation used, either \code{"mle"} or \code{"bayes"},}
\item{\code{chain}}{ -- if \code{method == "bayes"}, then this element will contain the Markov Chain from the sampler,
that is, hopefully a sample from the posterior density of the mixing parameter and the shape parameter.
if \code{method == "mle"} then this element does not exist.}
}

The output can be used in a variety of ways. If the interest is just in the
mixing and shape parameter estimates, then the \code{pi} and \code{shape}
member of the \code{psFit} object contains this information. It is also
displayed along with a number of fitted probabilities by the
\code{\link{print.psFit}} method. The fitted object can also be plotted using
the plot method \code{\link{plot.psFit}}, and to create a probability
function with \code{\link{probfun}}. **NOTE** The value of the shape
parameter that is printed (if you print the fitted object) is different from
that value that is stored in \code{shape}. The stored value is for the
\pkg{VGAM} parameterisation of the zeta distribution which uses
\eqn{s^\prime = s - 1}{s' = s - 1}. Therefore the printed value is \eqn{s =
s^\prime + 1}{s = s' + 1}. If you intend to use the fitted value with
\code{\link[VGAM]{dzeta}}, then you should use the stored value
\eqn{s^\prime}{s'}.

This function implements both maximum likelihood estimation (MLE) and
Bayesian estimation. Both modes of estimation require addition information
such as starting values and parameters for priors. Please read the
documentation for the \code{...} argument closely because it explains what
you can change and what the default values are.

Currently the Bayesian estimation is done assuming a Beta(alpha, beta)
distribution for the prior of the mixing proportion and Uniform[a, b] prior
  for the logarithm of the shape parameter. That is we assume \eqn{\log(s^\prime)
  \sim U[a,b]}{log(s') ~ U[a,b]}. This may change to be more flexible in the
future. Similarly, the estimation is done using a simple Metropolis-Hastings
sampler. It might be more efficient to sample through adaptive rejection
sampling, but it is unclear whether it is worth the effort.
}
\examples{
data(Psurveys)
roux = Psurveys$roux
fit = fitZIDist(roux)
fit
}
\references{
Evett, I. W. and Buckleton, J. S., "The interpretation of glass
  evidence. A practical approach", Journal of the Forensic Science Society
  1990: 30(4): 215--223.
}
\seealso{
\code{\link{plot.psFit}}, \code{\link{print.psFit}},
  \code{\link{probfun}}.
}
