% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.psFit.R
\name{logLik.psFit}
\alias{logLik.psFit}
\title{Extract the log-likelihood from a zeta model fit}
\usage{
\method{logLik}{psFit}(object, ...)
}
\arguments{
\item{object}{An object of class `"psFit"`.}

\item{...}{Additional arguments passed to methods. Currently ignored.}
}
\value{
An object of class `"logLik"` with attributes `"df"` and `"nobs"`.
}
\description{
Returns the maximised log-likelihood for a fitted zeta model.
}
\details{
This method allows generic functions such as [stats::AIC()] and
[stats::BIC()] to work with objects of class `"psFit"`.
}
