% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePrior.R
\name{makePrior}
\alias{makePrior}
\title{Define a prior density}
\usage{
makePrior(family = c("loguniform", "uniform", "custom"), range, logd)
}
\arguments{
\item{family}{One of \code{"loguniform"}, \code{"uniform"} or \code{"custom"}.}

\item{range}{Optionally the range for which the prior density is
evaluated. It is zero outside of this range. Parameter is
required when \code{family="custom"}.}

\item{logd}{Optionally (required when \code{family="custom"}.) a function
that evaluates the log density of the prior inside the range.}
}
\value{
an object of type \code{psPrior}
}
\description{
Construct a prior that can be used in the \code{\link{fitdist}} function.
}
\details{
The default is a LogUniform[-2,2] prior.
}
\examples{
## With default parameters, the prior will be LogUniform[-2, 2]
p1 <- makePrior()

# plot the prior density
xPlot <- seq(from = 0, to = 10, length = 100)
plot(xPlot, exp(p1$logd(xPlot)), type="l")

# Alternatively, a Uniform[a, b] prior can be used
p2 <- makePrior(family = "uniform", range = c(1,10))
plot(xPlot, exp(p2$logd(xPlot)), type="l")

# A custom prior needs the log density function to be specified
# We define an exponential prior with rate = 1
logdexp <- function(x) dexp(x, rate = 1/10, log = TRUE)
p3 <- makePrior(family = "custom", range = c(1, 10), logd = logdexp)

plot(xPlot, exp(p3$logd(xPlot)), type="l")

}
\seealso{
readData
}
