% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psModel.R
\name{modelObservationData}
\alias{modelObservationData}
\title{Convert fitPS data to the observation support required by a model.}
\usage{
modelObservationData(model, x, ...)
}
\arguments{
\item{model}{A `psModel` object.}

\item{x}{An object of class `psData`.}

\item{...}{Additional arguments reserved for model methods.}
}
\value{
Model-specific observation data.
}
\description{
Convert fitPS data to the observation support required by a model.
}
