% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.psFit.R
\name{predict.psFit}
\alias{predict.psFit}
\title{S3 predict method for an object of class \code{psFit}}
\usage{
\method{predict}{psFit}(x, newdata, interval = c("none", "confidence"), level = 0.95, ...)
}
\arguments{
\item{x}{an object of class \code{psFit}, usually from \code{\link{fitDist}}}

\item{newdata}{an optional vector of integers at which to calculate
\eqn{\Pr(X = x)}{Pr(X = x)}}

\item{interval}{either \code{"none"} or \code{"confidence"}, can be
abbreviated. If \code{"confidence"} then the bounds of a 100 * \code{level}
confidence interval for each predicted probability is provided.}

\item{level}{the level of a confidence interval. Ignored if
\code{interval == "none"}.}

\item{...}{other arguments passed to \code{predict}---not used}
}
\value{
either a named vector of fitted probabilities, or a \code{data.frame}
with columns \code{predicted}, \code{lower}, and \code{upper} and the row names
set to show what terms are being calculatd
}
\description{
S3 predict method for an object of class \code{psFit}
}
\examples{
data(Psurveys)
roux = Psurveys$roux
fit = fitDist(roux)
predict(fit, interval = "conf")
}
