% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBootstrap.R
\name{print.psBootstrap}
\alias{print.psBootstrap}
\title{Print a fitPS bootstrap object}
\usage{
\method{print}{psBootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class `psBootstrap`.}

\item{...}{Additional arguments passed to `print.data.frame()`.}
}
\value{
The bootstrap object, invisibly.
}
\description{
Print a fitPS bootstrap object
}
