% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.psFit.R
\name{print.psFit}
\alias{print.psFit}
\title{S3 print method for an object of class \code{psFit}}
\usage{
\method{print}{psFit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{psFit}, usually from \code{\link{fitDist}}
or #' \code{\link{fitZIDist}}.}

\item{...}{other arguments passed to \code{print}.}
}
\value{
No return value, called for side effects.
}
\description{
S3 print method for an object of class \code{psFit}
}
