% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probfun.R
\name{probfun}
\alias{probfun}
\title{Probability Functions}
\usage{
probfun(psFitobj)
}
\arguments{
\item{psFitobj}{an object of class \code{psFit}--see \code{\link{fitDist}}
and \code{\link{fitZIDist}}.}
}
\value{
a function that can be used to calculate any P or S term.
}
\description{
Creates a probability function that allows the computation of any P or S
term.
}
\examples{
p = readData(system.file("extdata", "p.xlsx", package = "fitPS"))
fit = fitDist(p)
P = probfun(fit)
P(0:5)
}
