% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzeta.R
\name{rzeta}
\alias{rzeta}
\title{Generate random variates from a zeta distribution}
\usage{
rzeta(n, shape)
}
\arguments{
\item{n}{Same as \code{\link[stats]{Poisson}}.}

\item{shape}{The standard zeta shape parameter, greater than 1.

See \code{\link[VGAM]{rzeta}}.}
}
\description{
Generate random variates from a zeta distribution
}
