% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZIzeta.R
\name{rZIzeta}
\alias{rZIzeta}
\title{Generate zero inflated zeta random variates}
\usage{
rZIzeta(n, pi = 0.5, shape = 1)
}
\arguments{
\item{n}{the number of observations.}

\item{pi}{the mixing parameter for the zero-inflated zeta model---must be in
(0, 1).}

\item{shape}{the shape parameter for the zero-inflated zeta. Must be greater
than zero.}
}
\value{
a vector of random variates from a zero-inflated zeta model
}
\description{
Generate zero inflated zeta random variates
}
\examples{
data(Psurveys)
roux = Psurveys$roux
fit.zi = fitZIDist(roux)
x = rZIzeta(n = sum(roux$data$rn), pi = fit.zi$pi, shape = fit.zi$shape)
table(x)
}
