% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.psFit.R
\name{summary.psFit}
\alias{summary.psFit}
\title{S3 summary method for an object of class \code{psFit}}
\usage{
\method{summary}{psFit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{psFit}, usually from \code{\link{fitDist}}
or \code{\link{fitZIDist}}}

\item{...}{other arguments passed to \code{summary}}
}
\value{
No return value, called for side effects
}
\description{
S3 summary method for an object of class \code{psFit}
}
\details{
Experimental because I am unsure if it is useful. If \code{object} is a zero-inflated zeta fitted object,
then the function carries out a likelihood ratio test for the value of pi. Currently not implemented for the logarithmic
distribution because we are currently not interested in the logarithmic distribution.
}
