% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.psData.R
\name{var}
\alias{var}
\title{Variance generic}
\usage{
var(x, ...)
}
\arguments{
\item{x}{an object for which we want to compute the sample variance.}

\item{\dots}{Any additional arguments to be passed to \code{var}.}
}
\description{
Variance generic
}
